+++ utils/__init__.py
@@ -1,2 +1,2 @@
 # Utils Package
-# Utilidades y Helpers para VentaPro
+# Utilidades y Helpers para NegocioSmart